requirements.txt
djangorestframework==3.7.7
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.7
whitenoise==3.2
